% Generated by roxygen2 (4.0.2): do not edit by hand
\name{all.equal.mult}
\alias{all.equal.mult}
\title{Compares >=2 objects with R function all.equal()}
\usage{
\method{all}{equal.mult}(...)
}
\arguments{
\item{...}{>=2 comma-separated object names}
}
\value{
\code{all.equal.mult} returns TRUE or list of pairwise all.equal() object comparisons
}
\description{
Compares >=2 objects with R function all.equal()
}
\examples{
foo <- c(1:10)
 bar <- c(1:10)
 foz <- c(1:10)
 baz <- letters[1:10]

 all.equal.mult(foo, bar) # TRUE
 all.equal.mult(foo, baz) # results of all.equal(foo, baz) as one-item list
 all.equal.mult(foo, bar, foz) # TRUE
 all.equal.mult(foo, bar, baz) # list of pairwise all.equal() comparisons among objects
}

