% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_equal_mult.R
\name{all_equal_mult}
\alias{all_equal_mult}
\title{Compares >=2 objects with R function all.equal()}
\usage{
all_equal_mult(...)
}
\arguments{
\item{...}{>=2 comma-separated object names}
}
\value{
\code{all_equal_mult} returns TRUE or list of pairwise all.equal() object comparisons
}
\description{
Compares >=2 objects with R function all.equal()
}
\examples{
 foo <- c(1:10)
 bar <- c(1:10)
 foz <- c(1:10)
 baz <- letters[1:10]

 all_equal_mult(foo, bar) # TRUE
 all_equal_mult(foo, baz) # results of all.equal(foo, baz) as one-item list
 all_equal_mult(foo, bar, foz) # TRUE
 all_equal_mult(foo, bar, baz) # list of pairwise all.equal() comparisons among objects
}

