% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_bead.R
\name{generate_bead}
\alias{generate_bead}
\title{Generate dataset for vignettes and simulation studies}
\usage{
generate_bead()
}
\value{
\code{\link[tibble]{tibble}} data frame
}
\description{
Generate dataset for vignettes and simulation studies
}
\examples{
set.seed(23)
df <- generate_bead()
df
}
