% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_real.R
\name{generate_real}
\alias{generate_real}
\title{Generate dataset for vignettes and simulation studies}
\usage{
generate_real()
}
\value{
\code{\link[tibble]{tibble}} data frame
}
\description{
Generate dataset for vignettes and simulation studies
}
\examples{
set.seed(23)
df <- generate_real()
df
}
