% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse_labels.R
\name{glimpse_labels}
\alias{glimpse_labels}
\title{Get a glimpse of your data, including variable labels.}
\usage{
glimpse_labels(x, width = NULL, labels = TRUE)
}
\arguments{
\item{x}{An object to glimpse at.}

\item{width}{Width of output: defaults to the setting of the option
`tibble.width` (if finite) or the width of the console.}

\item{labels}{Include variable labels in the printout. If `TRUE` and no 
variables have the `label` attribute, labels will be printed as 10
spaces to demonstrate their absence.}
}
\value{
x original x is (invisibly) returned, allowing \code{glimpse_labels()}
  to be used within a data pipe line.
}
\description{
This is like a transposed version of print: columns run down the page,
and data runs across. This makes it possible to see every column in
a data frame. It's a little like \code{str()} applied to a data frame
but it tries to show you as much data as possible. (And it always shows
the underlying data, even when applied to a remote data source.)
}
\details{
This is a modified version of the tibble package's 'glimpse.tbl' S3 method 
that includes the contents of the 'label' variable attribute, if present
(e.g., if you imported a Stata dataset with variable labels using 
\code{haven::read_stata()}).
}
\examples{
 dat1 <- tibble::data_frame(x = 1:20, y = 21:40)
 attributes(dat1[["x"]])$label <- "this is my label"
 glimpse_labels(dat1)
 
 dat2 <- dat1
 attributes(dat2[["x"]])$label <- paste0(rep("my label", 12), collapse = " ")
 glimpse_labels(dat2)
 
 dat3 <- dat1
 attributes(dat3[["x"]])$label <- paste0(rep("my label", 15), collapse = " ")
 glimpse_labels(dat3)
}

