% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compCytof.R
\name{compCytof}
\alias{compCytof}
\title{Compute spillover probability and correct for spillover}
\usage{
compCytof(sce, sce_bead, marker_to_barc, overwrite = FALSE)
}
\arguments{
\item{sce}{Single Cell Experiment for the real cells}

\item{sce_bead}{Single Cell Experiment for the bead experiment}

\item{marker_to_barc}{Table that maps the marker to the barcode in the beads experiment}
}
\value{
A list of class \code{spillr} containing
  \item{tb_compensate}{corrected real cells}
  \item{tb_spill_prob}{probability curve}
  \item{target_marker}{input marker in real experiment}
}
\description{
Compute spillover probability and correct for spillover
}
\examples{
sce <- prepData(mp_cells)    # real cells
sce_bead <- prepData(ss_exp) # beads
sce <- spillR::compCytof(sce, sce_bead, overwrite = FALSE)
}
