% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compCytof.R
\name{compCytof}
\alias{compCytof}
\title{Compute spillover probability and correct for spillover}
\usage{
compCytof(sce, sce_bead, marker_to_barc, overwrite = FALSE)
}
\arguments{
\item{sce}{Single Cell Experiment for the real cells}

\item{sce_bead}{Single Cell Experiment for the bead experiment}

\item{marker_to_barc}{Table for barcodes in the beads experiment}

\item{overwrite}{logical; if TRUE data are overwritten if FALSE 
data are saved in new columns}
}
\value{
Compensates the input \code{\link{SummarizedExperiment}} or, 
    if \code{x} is a character string, 
    all FCS files in the specified location.
}
\description{
Compute spillover probability and correct for spillover
}
\examples{
library(CATALYST)
library(dplyr)
library(magrittr)
bc_key <- c(139, 141:156, 158:176)
sce_bead <- prepData(ss_exp)
sce_bead <- assignPrelim(sce_bead, bc_key, verbose = FALSE)
sce_bead <- applyCutoffs(estCutoffs(sce_bead))
sce_bead <- computeSpillmat(sce_bead)
data(mp_cells, package = "CATALYST")
sce <- prepData(mp_cells)
marker_to_barc <- rowData(sce_bead)[,c("channel_name", "is_bc")] \%>\%
    as_tibble \%>\%
    dplyr::filter(is_bc == TRUE) \%>\%
    mutate(barcode = bc_key) \%>\%
    dplyr::select(marker = channel_name, barcode)
spillR::compCytof(sce, sce_bead, marker_to_barc, overwrite = FALSE)
}
