% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiagnostics.R
\name{plotDiagnostics}
\alias{plotDiagnostics}
\title{Compute spillover probability and correct for spillover}
\usage{
plotDiagnostics(sce, ch)
}
\arguments{
\item{sce}{A \code{\link[SingleCellExperiment]{SingleCellExperiment}} object}

\item{ch}{Character string specifying the channel to plot}
}
\value{
A list of \code{\link[ggplot2]{ggplot2}} plots
}
\description{
Compute spillover probability and correct for spillover
}
